// src/pages/MyOrders/models/DataType.tsx
export interface DataType {
  preOrderId: number;
  userIdDn: number;
  cpfCnpj: string;
  name: string;
  status: string;
  total: number;
  number: string;
  personType: string;
  statusDescription: string;
  notes: string;
  dateCreated: string;
  requestIdentification: string;
  contatoNome: string;
  contatoCPF: string;
  contatoPhone: string;
  contatoEmail: string;
}